Fixed Build
Fixed clippy errors in order-service

File: order_handlers.rs
Removed unnecessary cast:
count: r.count.unwrap_or(0) as i64 → count: r.count.unwrap_or(0)
Removed redundant cast in exchange math:
(row.unit_price_cents as i64).saturating_mul(it.qty as i64) → row.unit_price_cents.saturating_mul(it.qty as i64)
Fixed clippy errors in customer-service

File: main.rs
Replaced as_ref().map(|v| v.as_slice()) with as_deref() in four .bind(...) calls to satisfy option-as-ref-deref:
.bind(email_encrypted_param.as_deref())
.bind(phone_encrypted_param.as_deref())
.bind(email_hash_param.as_deref())
.bind(phone_hash_param.as_deref())
Verified clippy is clean locally

Ran clippy across the entire workspace: PASS (warnings only about future incompatibility in deps).
Improved Rust CI workflow

File: rust-ci.yml
Switched to dtolnay/rust-toolchain@stable and actions/cache@v4.
Ensured all cargo commands use working-directory: services.
Added a non-blocking “future incompatibility report” step to surface upstream warnings in CI without failing builds.
Minor workflow consistency tweaks

backend.yml: cache action v4, confirmed working directory.
sqlx-prepare-check.yml: fixed indentation in the version parity step to avoid shell parsing gotchas.
Kept integration-tests.yml as-is (already uses explicit --manifest-path services/Cargo.toml).
Validation runs
Build + clippy locally

cargo build --workspace --all-targets --locked: PASS
cargo clippy --workspace --all-targets -- -D warnings: PASS (no errors)
Prior frontend E2Es are green (admin and POS). I also fixed the POS exchange E2E earlier by seeding the session pre-navigation and adding TS annotations.

+++ services/order-service/src/order_handlers.rs
@@ -323,7 +323,7 @@ pub async fn get_settlement_report(
     for r in rows {
         totals.push(SettlementByMethod {
             method: r.method,
-            count: r.count.unwrap_or(0) as i64,
+            count: r.count.unwrap_or(0),
             amount: r.amount.unwrap_or(BigDecimal::from(0)),
         });
     }
@@ -428,7 +428,7 @@ pub async fn exchange_order(
         }
         for it in &req.return_items {
             let row = by_product.get(&it.product_id).unwrap();
-            let line_cents = (row.unit_price_cents as i64).saturating_mul(it.qty as i64);
+            let line_cents = row.unit_price_cents.saturating_mul(it.qty as i64);
             refund_total_cents = refund_total_cents.saturating_add(line_cents);
             // update returned_quantity and insert return item row
             let conn = tx.acquire().await.map_err(|e| ApiError::Internal { trace_id: None, message: Some(format!("Failed to acquire conn: {}", e)) })?;
